[Changes] CHexCtrl friend class dependency removed/reworked. CHexCtrl private methods to public. CHexBookmarks reworked. Search routines moved to CHexDlgSearch and drastically reworked. CScrollEx to give the focus on LMouseClick. [Added] CHexDlgSearch::DoSearch, CHexCtrl::GetData(), OnKeyDownShiftLeft, OnKeyDownShiftRight, OnKeyDownShiftUp, OnKeyDownShiftDown. [Fixes] ReplaceAll UTF16 bug. OnKeyDownShiftUp selection inconsistency fix. Differetnt m_iSpaceBetweenBlocks calc for when capacity == 1. Other small fixes.

+++ HexCtrl/src/Dialogs/CHexDlgFillWith.cpp
@@ -87,7 +87,7 @@ void CHexDlgFillWith::OnOK()
 		break;
 	case IDC_HEXCTRL_FILLWITHDATA_RADIO_UTF16:
 		hms.pData = (PBYTE)wstrComboText.data();
-		hms.ullDataSize = wstrComboText.size() * sizeof(WCHAR);
+		hms.ullDataSize = (ULONGLONG)wstrComboText.size() * sizeof(WCHAR);
 		break;
 	}
